refactor: uplift list summary card
