+++ projects/client/src/lib/sections/lists/components/list-summary/ListSummaryCard.svelte
@@ -2,8 +2,8 @@
   import Card from "$lib/components/card/Card.svelte";
   import type { MediaListSummary } from "$lib/requests/models/MediaListSummary.ts";
   import type { MediaType } from "$lib/requests/models/MediaType.ts";
-  import ListHeader from "./ListHeader.svelte";
-  import ListPosters from "./ListPosters.svelte";
+  import ListHeader from "./_internal/ListHeader.svelte";
+  import ListPosters from "./_internal/ListPosters.svelte";
 
   const { list, type }: { list: MediaListSummary; type: MediaType } = $props();
 </script>
