+++ projects/client/src/lib/sections/lists/components/list-summary/_internal/ListPosters.svelte
